+++ packages/rfw/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.26
+* Supports overriding the error widget builder.
+
+## 1.0.25
+* Adds support for wildget builders.
+
 ## 1.0.24
 
 * Adds `InkResponse` material widget.